feat: Save duplicate events (#15044)
This PR switches from a first write wins to a last write wins in the case
of events with duplicate IDs.

This is the last place that we read from the Postgres events table. Since
this will be removed soon, we won't be able to check this for the
presence of a duplicate event ID anymore.